fix(codex): keep usage parser plugin-local
