+++ src/infra/provider-usage.fetch.ts
@@ -1,6 +1,6 @@
 // Public facade for built-in provider usage fetch implementations.
 export { fetchClaudeUsage } from "./provider-usage.fetch.claude.js";
-export { buildCodexAppServerUsageSnapshot, fetchCodexUsage } from "./provider-usage.fetch.codex.js";
+export { fetchCodexUsage } from "./provider-usage.fetch.codex.js";
 export { fetchDeepSeekUsage } from "./provider-usage.fetch.deepseek.js";
 export { fetchGeminiUsage } from "./provider-usage.fetch.gemini.js";
 export { fetchMinimaxUsage } from "./provider-usage.fetch.minimax.js";
